Fix taplo test target directories (#36690)
The command `./mach test-tidy` did not check the TOML files in the
subdirectories of components/shared/, since the argument for taglo
command was written as `components/shared/*.toml`.

This patch fixes it to include those TOML files in test-tidy.

Testing: It only involves python script for test-tidy. No test is
needed.
Fixes: #36659

Signed-off-by: Kingsley Yung <kingsley@kkoyung.dev>