net: Fix unit tests requiring default resources. (#36838)
When all unit tests are run, the existing feature flag use from
net_traits was enough to ensure the default resources are baked into the
test binary. When only the unit tests for the `net` crate are run, that
feature flag was not enabled in the build so the tests that require the
resources fail.

Testing: Ran `./mach test-unit -p net`
Fixes: #36837

Signed-off-by: Josh Matthews <josh@joshmatthews.net>