constellation: Ignore `zombie_process` warning (#36904)
`rust-analyzer` shows a `zombie_process` warning, but only on macOS as
far
as I can tell. This change ignores the warning here. We aren't currently
in a position where we can reliably `wait()` on spawned processes, which
needs to happen as part of a larger multiprocess / cleanup effort.

Testing: No testing as this just ignores an LSP warning.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>