webgl: Use `glow::Context::supported_extensions()` to implement `getSupportedExtensions()` (#36911)
Not only does this simplify the code, it fixes a problem where we were
attempting to use an OpenGL 3.0 API on an incompatible GL context.

Testing: There are already tests for `getSupportedExtensions()` in the
WebGL
suite, but effectively testing this requires drivers that do not support
a particular version of OpenGL, so it is a bit hard to actually test.
Fixes: #36852.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>