bootstrap: Add `winget` fallback (#32836)
When `choco` is not available, we can install the same packages from
winget. winget is an official package manager from Microsoft, which is
available by default on Windows 11.

Note: `winget` also has non-interactive installation options, but
accepting license agreements should still be the responsibility of the
user, so we don't add any such option for now.

---

- [x] `./mach build -d` does not report any errors
- [x] `./mach test-tidy` does not report any errors
- [x] These changes fix: Bootstrap on windows without `choco` available.

Signed-off-by: Jonathan Schwender <schwenderjonathan@gmail.com>