Add alt-d shortcut to select address bar text (#37794)
I usually use the alt-d in shortcut in Firefox to select the text in the
address bar. Firefox supports both alt-d and ctrl-l. This change adds
the alt-d shortcut so you can use either.

Testing: I tested it by running Servo and pressing alt-d. I don't know
if there is a way to create a test for it?

---------

Signed-off-by: Michael Mc Donnell <michael@mcdonnell.dk>