servoshell: Select address bar text on click (#37839)
In Firefox, clicking on the address bar selects all the text. This makes
Servo to do the same. I reworked the code so that the shortcut only
changes the focus. That means that either clicking or the shortcut
changes the focus which in turn selects the text.

Testing: I tested it manually by clicking in the address bar and using
the shortcuts and noticing that the text was selected.

Signed-off-by: Michael Mc Donnell <michael@mcdonnell.dk>