WPT upstreamer: Make all failed merge operations leave a comment. (#37949)
The WPT exporter currently assumes that removing a label from a github
PR will never fail. This is obviously false and the result is a silent
failure. By moving the operation inside the existing try block we ensure
that a comment will be posted if any part of the merge step fails.

Testing: Untested; did not feel worth the time investment to determine
if the existing test harness can mock this particular situation.
Fixes: #37914

Signed-off-by: Josh Matthews <josh@joshmatthews.net>